Move MoreArrays to general utils
