+++ lenskit-core/src/test/java/org/lenskit/util/MoreArraysTest.java
@@ -18,12 +18,12 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.grouplens.lenskit.collections;
+package org.lenskit.util;
+
+import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.*;
-
-import org.junit.Test;
 
 /**
  * Test the array utility methods.
